### 2016-11-02 #104 @whedon
[PRE REVIEW]: resampy: efficient sample rate conversion in Python
**Submitting author:** bmcfee (<a href="http://orcid.org/0000-0001-6261-9747">Brian McFee</a>)
**Repository:** <a href="https://github.com/bmcfee/resampy" target ="_blank">https://github.com/bmcfee/resampy</a>
**Version:** 0.1.4

Thanks for submitting your paper to JOSS bmcfee. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
