### 2016-11-28 #125 @whedon
[REVIEW]: resampy: efficient sample rate conversion in Python
**Submitting author:** @bmcfee (<a href="http://orcid.org/0000-0001-6261-9747">Brian McFee</a>)
**Repository:** <a href="https://github.com/bmcfee/resampy" target ="_blank">https://github.com/bmcfee/resampy</a>
**Version:** 0.1.4

- [x] **Authorship:** Has the submitting author (bmcfee) made major contributions to the software?
