### 2016-11-29 bmcfee/resampy#37 @arokem
[JOSS review] Installation instructions in sphinx docs
Would be good to move/copy the installation instructions into the sphinx documentation. Might be good to add instructions on building the software from source.
