### 2016-11-29 bmcfee/resampy#39 @arokem
[JOSS review] Contribution guidelines
"Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

Seems to me that item 2 is covered through the link to the issue tracker, but it would be good to (at the very least) add a `CONTRIBUTING` file to the repo (see: https://github.com/blog/1184-contributing-guidelines).
