### 2016-11-29 bmcfee/resampy#40 @arokem
[JOSS Review] References require a DOI
Required by JOSS review criteria:

"Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?"
