### 2016-12-02 bmcfee/resampy#38 @bmcfee
Thanks for the review/comments @arokem!  Some responses:

> Why are tests in the top-level directory, rather than under the package?

I suppose this is a habit I picked up from developing packages with large regression test dependencies.  More generally, I prefer to keep tests separate from the package so that the tests operate in an environment that best matches application code (ie, package code is loaded by absolute import instead of relative).

This issue came up previously in building deb packages #36.  I'm open to moving the tests in this package, but haven't yet done so.

> Having said that, even after building the package, I get the following test errors (maybe I'm doing it wrong?):

The tests are written for `py.test`, not `nosetests`, so that probably explains the errors you're getting.

I guess this means the documentation should be expanded to include running tests. :grin:
