### 2016-12-02 bmcfee/resampy#38 @arokem
Gotcha.

Yes: documentation about testing would be good.

But your call about moving the tests. Feel free to close this issue.
