### 2016-12-02 bmcfee/resampy#41 @arokem
[JOSS Review] Performance claims
How would I confirm the performance claims of the software? An example that demonstrates the speedup benefits would be a useful addition to the documentation, IMHO.
