### 2016-12-02 bmcfee/resampy#38 @bmcfee
Okay.  I think I'll just update the documentation for now, and revisit test migration at the next release.  Thanks!
