### 2016-12-02 bmcfee/resampy#41 @bmcfee
Great idea.  I have this implemented in notebooks that could be readily pushed into the documentation.

Here's a quick example from the scipy lightning talk:

```python
import numpy as np
import scipy
import resampy

x = np.random.randn(1, 400000)
sr_in, sr_out = 22050, 16000

%timeit resampy.resample(x, sr_in, sr_out, axis=-1)

%timeit scipy.signal.resample(x, int(x.shape[1] * sr_out / sr_in), axis=-1)
```
