### 2016-12-02 bmcfee/resampy#40 @bmcfee
Are you referring to the package itself (eg zenodo DOI), or the source material reference (Smith) in the docs?

For the latter: he explicitly requests a specific form of citation to his web book, which does not currently have a DOI.  Any advice about how to handle that?
