### 2016-12-02 bmcfee/resampy#40 @arokem
I guess there's not much to do about the web-book (except maybe to ask the author to DOI it, if only so that it's available when he retires, or when his group's website is migrated...). Does the software referenced have a DOI?
