### 2016-12-02 bmcfee/resampy#41 @arokem
I am much more impressed now. Feel free to close this issue.
