### 2016-12-02 bmcfee/resampy#41 @arokem
By the way, is there a video of this lightning talk somewhere? You might want to add that to the docs as well.
