### 2016-12-02 bmcfee/resampy#40 @bmcfee
Yes, there's a zenodo DOI.  The badge is in the updated readme in #42 .
