### 2016-12-02 513780b
added benchmark to documentation
