### 2016-12-02 bmcfee/resampy#41 @bmcfee
Benchmark example is now up in the #42 docs.
