### 2016-12-02 bmcfee/resampy#40 @arokem
To be clear -- this reference:

https://github.com/bmcfee/resampy/blob/master/paper/paper.bib#L9

still requires a DOI, if such exists
