### 2016-12-03 bmcfee/resampy#40 @bmcfee
libsamplerate also does not appear to have a DOI.
