### 2016-12-03 65e759d
switched miniconda to miniconda3
