+++ .travis.yml
@@ -26,7 +26,7 @@ before_install:
 - export PATH="$MINICONDA/bin:$PATH"
 - hash -r
 # Install conda only if necessary
-- command -v conda >/dev/null || { wget http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh -O miniconda.sh;
+- command -v conda >/dev/null || { wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
    bash miniconda.sh -b -f -p $MINICONDA; }
 - conda config --set always_yes yes
 - conda update conda
