### 2016-12-03 bmcfee/resampy#37 @bmcfee
Fixed by merging #42
