### 2016-12-03 bmcfee/resampy#38 @bmcfee
Fixed by merging #42
