### 2016-12-03 bmcfee/resampy#39 @bmcfee
Fixed by merging #42
