### 2016-12-03 bmcfee/resampy#41 @bmcfee
Fixed by merging #42
