### 2016-12-03 bmcfee/resampy#40 @arokem
OK. I guess that's not too important, because who would use that, now that
they have resampy? :-)

Closing.

On Sat, Dec 3, 2016 at 9:40 AM, Brian McFee <notifications@github.com>
wrote:

> libsamplerate also does not appear to have a DOI.
>
