### 2016-12-03 #125 @arokem
👍 from me for publication.

See:

https://github.com/bmcfee/resampy/issues/37
https://github.com/bmcfee/resampy/issues/38
https://github.com/bmcfee/resampy/issues/39
https://github.com/bmcfee/resampy/issues/40
https://github.com/bmcfee/resampy/issues/41
