### 2016-12-05 #125 @arfon
@arokem many thanks for the review here!

@bmcfee - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00125 🚀 🎉 💥
